backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
groq==1.0.0
anthropic==0.7.8
langchain==0.1.0
langchain-openai==0.0.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
pytz==2023.3.post1
aiofiles==23.2.1
httpx==0.25.2

# Authentication & Security
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Payments (Stripe - Optional but Recommended)
stripe==7.8.0